+++ css/custom-style.css
@@ -1,0 +1,5 @@
+.loadingerror {
+  max-width: 800px;
+  margin: 0 auto;
+  background: #f04946;
+}
