+++ css/style.css
@@ -1336,6 +1336,16 @@ main {
 }
 
 
+
+@media (min-width: 1024px) {
+  .img-upload__wrapper {
+    padding-bottom: 0;
+  }
+}
+
+
+
+
 /* Адаптивности чуток
    ========================================================================== */
