+++ js/templates/error-template.js
@@ -1,0 +1,9 @@
+import {Template} from './template.js';
+
+
+const ERROR_HTML =
+;
+const errorTemplate = new Template(ERROR_HTML);
+
+
+export {errorTemplate};
